Implemented basic typing test functionality
Text box can now take in input using an invisible input. The text box displays a list of words, and updates
the color of the individual characters and a WPM counter based on the user input. Backspaces work within the
current word begin typed. It does not check for accuracy.