Made it so user info can be stored to any user after they log in
Still don't have the code for actually inputting the data so the template has been added to the top of server.js. The main thing is to always put in withCredentials: true since we are using the credentials to put the data in the correct account. Next step is to implement a state for wpm and accuracy so the actual stats can be put in the user's database in TypingTest.js and also making it so that the user can't go back to the login page after they have logged in and displaying this info in the profile page